+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.17.0] - 2023-02-01
+### Fixed
+- Fixed an edge case of missing very first change in a file from bufferHandler
+- bufferHandler updated to iterate new_files first before normal diffs
+
 ## [3.16.0] - 2023-01-30
 ### Updated
 - Added source=vscode as query parameters in API calls